Added Player (and Character) map object

+++ core/src/com/celtican/pokemon/utils/graphics/Renderable.java
@@ -1,5 +1,5 @@
 package com.celtican.pokemon.utils.graphics;
 
 public interface Renderable {
-    void render(float x, float y);
+    void render(int x, int y);
 }
